Remove any validation from the validator, its all done in rule now. (#1165)
* Merkle block rule

* Remove coinbase checks from validator

* Remove signature operation count rule

* Pow Transaction rule

* Remove POS check transaction

* Pos Block signature rule

* Remove the pos block checks

* Remove any validation from the validator, its all done in rule now

* Passing the logger to the transaction rules from the mempool validator

* Act on comments

* Fixing some spelling and comments

* Act on review
