+++ src/Stratis.Bitcoin.Features.Consensus/ConsensusFeature.cs
@@ -264,7 +264,7 @@ namespace Stratis.Bitcoin.Features.Consensus
                     // rules that are inside the method CheckBlock
                     new BlockMerkleRootRule(),
                     new EnsureCoinbaseRule(),
-                    new CheckTransactionRule(),
+                    new CheckPowTransactionRule(),
                     new CheckSigOpsRule()
                 };
             }
@@ -298,8 +298,12 @@ namespace Stratis.Bitcoin.Features.Consensus
                     // rules that are inside the method CheckBlock
                     new BlockMerkleRootRule(),
                     new EnsureCoinbaseRule(),
-                    new CheckTransactionRule(),
-                    new CheckSigOpsRule()
+                    new CheckPowTransactionRule(),
+                    new CheckPosTransactionRule(),
+                    new CheckSigOpsRule(),
+                    new PosFutureDriftRule(),
+                    new PosCoinstakeRule(),
+                    new PosBlockSignatureRule()
                 };
             }
         }
