+++ src/Stratis.Bitcoin.Features.Consensus/Rules/CommonRules/CheckPosTransactionRule.cs
@@ -1,0 +1,35 @@
+using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+using NBitcoin;
+
+namespace Stratis.Bitcoin.Features.Consensus.Rules.CommonRules
+{    /// <summary>
+     /// Validate a PoS transaction.
+     /// </summary>
+    public class CheckPosTransactionRule : ConsensusRule
+    {
+        /// <inheritdoc />
+        public override Task RunAsync(RuleContext context)
+        {
+            Block block = context.BlockValidationContext.Block;
+
+            // Check transactions
+            foreach (Transaction tx in block.Transactions)
+                this.CheckTransaction(tx);
+
+            return Task.CompletedTask;
+        }
+
+        public virtual void CheckTransaction(Transaction transaction)
+        {
+            foreach (TxOut txout in transaction.Outputs)
+            {
+                if (txout.IsEmpty && !transaction.IsCoinBase && !transaction.IsCoinStake)
+                {
+                    this.Logger.LogTrace("(-)[USER_TXOUT_EMPTY]");
+                    ConsensusErrors.BadTransactionEmptyOutput.Throw();
+                }
+            }
+        }
+    }
+}
