+++ src/Stratis.Bitcoin.Features.MemoryPool/MempoolValidator.cs
@@ -10,6 +10,7 @@ using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Features.Consensus;
 using Stratis.Bitcoin.Features.Consensus.CoinViews;
 using Stratis.Bitcoin.Features.Consensus.Interfaces;
+using Stratis.Bitcoin.Features.Consensus.Rules.CommonRules;
 using Stratis.Bitcoin.Features.MemoryPool.Interfaces;
 using Stratis.Bitcoin.Utilities;
 
@@ -536,8 +537,10 @@ namespace Stratis.Bitcoin.Features.MemoryPool
         /// <param name="context">Current validation context.</param>
         private void PreMempoolChecks(MempoolValidationContext context)
         {
-            // state filled in by CheckTransaction
-            this.consensusValidator.CheckTransaction(context.Transaction);
+            // TODO: fix this to use dedicated mempool rules.
+            new CheckPowTransactionRule { Logger = this.logger }.CheckTransaction(this.ConsensusOptions, context.Transaction);
+            if(this.chain.Network.NetworkOptions.IsProofOfStake)
+                new CheckPosTransactionRule { Logger = this.logger }.CheckTransaction(context.Transaction);
 
             // Coinbase is only valid in a block, not as a loose transaction
             if (context.Transaction.IsCoinBase)
