+++ src/Stratis.Bitcoin.IntegrationTests/CoinViewTests.cs
@@ -344,10 +344,11 @@ namespace Stratis.Bitcoin.IntegrationTests
             };
 
             Network.Main.Consensus.Options = new PowConsensusOptions();
+            context.Consensus = Network.Main.Consensus;
             ConsensusSettings consensusSettings = new ConsensusSettings().Load(NodeSettings.Default());
             var validator = new PowConsensusValidator(Network.Main, new Checkpoints(Network.Main, consensusSettings), DateTimeProvider.Default, this.loggerFactory);
             new WitnessCommitmentsRule().RunAsync(context).GetAwaiter().GetResult();
-            validator.CheckBlock(context);
+            new CheckPowTransactionRule().RunAsync(context);
         }
     }
 }
